+++ src/components/User/Assets/Issuance/components/IssuanceRequestList.vue
@@ -4,11 +4,18 @@
       <div class="issuance-rl__filters">
         <select-field
           class="issuance-rl__filter app-list-filters__field"
-          label="State" v-model="filters.state"
+          label="State"
+          v-model="filters.state"
         >
-          <option :value="REQUEST_STATES.pending">Pending</option>
-          <option :value="REQUEST_STATES.approved">Approved</option>
-          <option :value="REQUEST_STATES.permanentlyRejected">Permanently rejected</option>
+          <option :value="REQUEST_STATES.pending">
+            Pending
+          </option>
+          <option :value="REQUEST_STATES.approved">
+            Approved
+          </option>
+          <option :value="REQUEST_STATES.permanentlyRejected">
+            Permanently rejected
+          </option>
         </select-field>
         <select-field
           class="issuance-rl__filter app-list-filters__field"
@@ -28,10 +35,10 @@
 
     <div class="issuance-rl__request-counter-wrapper">
       <p class="issuance-rl__request-counter">
-        Pending requests: {{listCounter.pending}}
+        Pending requests: {{ listCounter.pending }}
       </p>
       <p class="issuance-rl__request-counter">
-        Fullfilled requests: {{listCounter.approved}}
+        Fullfilled requests: {{ listCounter.approved }}
       </p>
     </div>
     <div class="issuance-rl__list-wrp">
@@ -49,30 +56,34 @@
             <span class="app-list__cell">
               Requestor
             </span>
-
           </div>
-          <li v-for="item in list" :key="item.id" class="issuance-rl__li">
+          <li
+            v-for="item in list"
+            :key="item.id"
+            class="issuance-rl__li">
             <router-link
               class="issuance-rl__li-a"
               :to="{ name: 'assets.issuance.props', params: { id: item.id } }"
             >
-            <span
-              class="app-list__cell app-list__cell--important"
-              :title="`${localize(item.amount)} ${item.asset}`"
-            >
-              {{ localize(item.details.createIssuance.amount) }}
-              {{ item.details.createIssuance.asset }}
-            </span>
+              <span
+                class="app-list__cell app-list__cell--important"
+                :title="`${localize(item.amount)} ${item.asset}`"
+              >
+                {{ localize(item.details.createIssuance.amount) }}
+                {{ item.details.createIssuance.asset }}
+              </span>
 
-              <span class="app-list__cell app-list__cell--wrap"
-                    :title="item.createdAt">
-              {{item.createdAt | dateTime}}
-            </span>
+              <span
+                class="app-list__cell app-list__cell--wrap"
+                :title="item.createdAt">
+                {{ item.createdAt | dateTime }}
+              </span>
 
-              <span class="app-list__cell"
-                    :title="item.requestor">
-              {{item.requestor}}
-            </span>
+              <span
+                class="app-list__cell"
+                :title="item.requestor">
+                {{ item.requestor }}
+              </span>
             </router-link>
           </li>
         </ul>
@@ -81,7 +92,7 @@
       <template v-else>
         <div class="app-list">
           <div class="app-list__li-like app-list__li--no-shadow">
-            <p>{{isLoaded ? 'Nothing here yet' : 'Loading...'}}</p>
+            <p>{{ isLoaded ? 'Nothing here yet' : 'Loading...' }}</p>
           </div>
         </div>
       </template>
@@ -102,10 +113,10 @@
 import IssuanceMixin from './issuance.mixin'
 
 export default {
-  mixins: [IssuanceMixin]
+  mixins: [IssuanceMixin],
 }
 </script>
 
 <style scoped lang="scss">
-  @import "./issuance";
+@import "./issuance";
 </style>