+++ src/components/User/Trades/components/PriceChart/PriceChart.Renderer.vue
@@ -1,9 +1,10 @@
 <template>
-  <div class="price-chart"></div>
+  <div class="price-chart" />
 </template>
 
 <script>
 import { AssetPair } from '../../models/AssetPair'
+
 import * as d3Array from 'd3-array'
 import * as d3Selection from 'd3-selection'
 import * as d3Scale from 'd3-scale'
@@ -14,7 +15,17 @@ import * as d3Ease from 'd3-ease'
 import * as d3Format from 'd3-format'
 
 import moment from 'moment'
-const d3 = Object.assign({}, d3Array, d3Selection, d3Axis, d3Shape, d3Scale, d3Transition, d3Ease, d3Format)
+const d3 = Object.assign(
+  {},
+  d3Array,
+  d3Selection,
+  d3Axis,
+  d3Shape,
+  d3Scale,
+  d3Transition,
+  d3Ease,
+  d3Format
+)
 
 const CLASS_NAME = 'price-chart'
 const DECIMAL_PRECISION = 4
@@ -22,6 +33,13 @@ const LARGE_NUMBER_PRECISION = 2
 
 export default {
   name: CLASS_NAME,
+
+  props: {
+    priceHistory: { type: Object, required: true },
+    scale: { type: String, required: true },
+    assetPair: { type: String, required: true },
+  },
+
   data () {
     return {
       assetCode: '',
@@ -32,16 +50,20 @@ export default {
       xAxis: '',
       yAxis: '',
       animationDuration: 500,
-      isAnimating: false
+      isAnimating: false,
     }
   },
 
-  props: ['priceHistory', 'scale', 'assetPair'],
+  computed: {
+    data () {
+      return this.normalizeData(this.priceHistory[this.scale])
+    },
+  },
 
   watch: {
     'priceHistory' () { this.update() },
     'scale' () { this.update() },
-    'assetPair' () { this.update() }
+    'assetPair' () { this.update() },
   },
 
   mounted (...args) {
@@ -51,12 +73,6 @@ export default {
 
   beforeDestroy () {
     window.removeEventListener('resize', this.init)
-  },
-
-  computed: {
-    data () {
-      return this.normalizeData(this.priceHistory[this.scale])
-    }
   },
 
   methods: {
@@ -70,7 +86,7 @@ export default {
         width: parentElement.clientWidth,
         height: parentElement.clientHeight < 250
           ? 250
-          : parentElement.clientHeight
+          : parentElement.clientHeight,
       }
     },
 
@@ -82,8 +98,11 @@ export default {
     },
 
     getMaxAndMinDates (data) {
-      const max = data.reduce((acc, { time: cur }) => cur > acc ? cur : acc, 0)
-      const min = data.reduce((acc, { time: cur }) => cur < acc ? cur : acc, max)
+      const max = data
+        .reduce((acc, { time: cur }) => cur > acc ? cur : acc, 0)
+      const min = data
+        .reduce((acc, { time: cur }) => cur < acc ? cur : acc, max)
+
       return { max, min }
     },
 
@@ -97,11 +116,11 @@ export default {
 
     formatMoney (amount) {
       const symbol = ({
-        'USD': '$'
+        'USD': '$',
       })[this.assetCode]
 
       const moneyFormats = {
-        'en': ({ value, symbol }) => `${value} ${this.assetCode}`
+        'en': ({ value, symbol }) => `${value} ${this.assetCode}`,
       }
 
       return moneyFormats['en']({ value: this.prettifyNumber(amount), symbol })
@@ -111,7 +130,7 @@ export default {
       return data
         .map(item => ({
           time: moment(item.timestamp).toDate(),
-          value: +item.value
+          value: +item.value,
         }))
         .sort(function (a, b) { return a.time - b.time })
     },
@@ -130,7 +149,7 @@ export default {
       return [
         average - gap,
         average,
-        average + gap
+        average + gap,
       ]
     },
 
@@ -168,12 +187,12 @@ export default {
           new Date(average - gap * 3),
           new Date(average - gap),
           new Date(average + gap),
-          new Date(average + gap * 3)
+          new Date(average + gap * 3),
         ]
       } else {
         result = [
           new Date(average - gap * 3),
-          new Date(average + gap * 3)
+          new Date(average + gap * 3),
         ]
       }
 
@@ -221,7 +240,7 @@ export default {
         top: 20,
         right: 5,
         bottom: 30,
-        left: this.genYMaxTickWidth(this.data)
+        left: this.genYMaxTickWidth(this.data),
       }
       const dimensions = this.getDimensions(this.$el)
       this.width = dimensions.width - margin.right - margin.left
@@ -313,7 +332,7 @@ export default {
 
         this.isFirstRender = false
       } else {
-          // Update the line
+        // Update the line
         let path = this.svg.selectAll(`.${CLASS_NAME}__line`)
           .attr('stroke-dasharray', null)
           .attr('stroke-dashoffset', null)
@@ -413,8 +432,8 @@ export default {
         this.isAnimating = false
         priceChartSvg.classed(`${CLASS_NAME}--overflow-hidden`, false)
       }, this.animationDuration)
-    }
-  }
+    },
+  },
 }
 </script>