+++ src/components/User/Users/components/UserDetails/UserDetails.Kyc.vue
@@ -1,14 +1,20 @@
 <template>
   <div class="user-details-account">
-    <h3 class="user-details-account__title">KYC information</h3>
+    <h3 class="user-details-account__title">
+      KYC information
+    </h3>
     <div class="user-details-account__info">
       <p>Can be uploaded one of the documents:</p>
       <p> - Valid passport;</p>
       <p>- Driving license - full, not provisional, with a color photograph;</p>
-      <p> - Valid National ID card;
-        <p />
-        <p>- In all cases handwritten documents or those in a foreign script that cannot be read, i.e Arabic or Cyrillic
-          will not be accepted unless translated and notarized.</p>
+      <p>
+        - Valid National ID card;
+      </p><p />
+      <p>
+        - In all cases handwritten documents or those in a foreign script
+        that cannot be read, i.e Arabic or Cyrillic
+        will not be accepted unless translated and notarized.
+      </p>
     </div>
 
     <template v-if="isLoaded">
@@ -47,15 +53,21 @@
       <ul class="key-value-list">
         <li>
           <span>First name</span>
-          <span :title="kyc.firstName">{{kyc.firstName}}</span>
+          <span :title="kyc.firstName">
+            {{ kyc.firstName }}
+          </span>
         </li>
         <li>
           <span>Last name</span>
-          <span :title="kyc.lastName">{{kyc.lastName}}</span>
+          <span :title="kyc.lastName">
+            {{ kyc.lastName }}
+          </span>
         </li>
         <li>
           <span>ID Document type</span>
-          <span :title="ID_DOCUMENTS_VERBOSE[kyc.idDocumentType]">{{ID_DOCUMENTS_VERBOSE[kyc.idDocumentType]}}</span>
+          <span :title="ID_DOCUMENTS_VERBOSE[kyc.idDocumentType]">
+            {{ ID_DOCUMENTS_VERBOSE[kyc.idDocumentType] }}
+          </span>
         </li>
         <li>
           <span>ID Document</span>
@@ -106,16 +118,20 @@
         <p>Can be uploaded one of the documents:</p>
         <p> - Driving license if not already uploaded on the previous step;</p>
         <p>- Recent utility bill (not older than 6 months old);</p>
-        <p> - Bank statement or credit card statement (not older than 6 months);
-          <p />
-          <p> - Bank reference letter (not older than 6 months);
-            <p />
-            <p>- Signed Tenancy/Rental Agreement.</p>
+        <p>
+          - Bank statement or credit card statement (not older than 6 months);
+        </p><p />
+        <p>
+          - Bank reference letter (not older than 6 months);
+        </p><p />
+        <p>- Signed Tenancy/Rental Agreement.</p>
       </div>
     </template>
 
     <template v-else-if="isFailed">
-      <p class="danger">An error occurred. Please try again later.</p>
+      <p class="danger">
+        An error occurred. Please try again later.
+      </p>
     </template>
 
     <template v-else>
@@ -125,9 +141,11 @@
 </template>
 
 <script>
-import { UserDocGetter, UserDocLinkGetter, ImgGetter } from '@comcom/getters'
+import { UserDocGetter, UserDocLinkGetter } from '@comcom/getters'
+
 import { fromKycTemplate } from '../../../../../utils/kyc-tempater'
 import deepCamelCase from 'camelcase-keys-deep'
+
 import config from '@/config'
 import { Sdk } from '@/sdk'
 import { ErrorHandler } from '@/utils/ErrorHandler'
@@ -136,14 +154,18 @@ const ID_DOCUMENTS_VERBOSE = {
   passport: 'Passport',
   identity_card: 'Identity card',
   driving_license: 'Driving license',
-  residence_permit: 'Residence permit'
+  residence_permit: 'Residence permit',
 }
 
 export default {
   components: {
     UserDocGetter,
     UserDocLinkGetter,
-    ImgGetter
+  },
+
+  props: {
+    user: { type: Object, required: true },
+    blobId: { type: String, required: true },
   },
 
   data () {
@@ -152,11 +174,14 @@ export default {
       isLoaded: false,
       isFailed: false,
       config,
-      ID_DOCUMENTS_VERBOSE
+      ID_DOCUMENTS_VERBOSE,
     }
   },
 
-  props: ['user', 'blobId'],
+  watch: {
+    user () { this.getKyc() },
+    blobId () { this.getKyc() },
+  },
 
   created () {
     this.getKyc()
@@ -170,18 +195,16 @@ export default {
       try {
         const response = await Sdk.api.blobs.get(this.blobId, this.user.address)
         const kycFormResponse = response.data
-        this.kyc = deepCamelCase(fromKycTemplate(JSON.parse(kycFormResponse.value)))
+        this.kyc = deepCamelCase(
+          fromKycTemplate(JSON.parse(kycFormResponse.value))
+        )
         this.isLoaded = true
       } catch (error) {
-        ErrorHandler.process(error)
+        ErrorHandler.processWithoutFeedback(error)
         this.isFailed = true
       }
-    }
+    },
   },
-  watch: {
-    user () { this.getKyc() },
-    blobId () { this.getKyc() }
-  }
 }
 </script>