+++ src/components/common/getters/EmailGetter.vue
@@ -27,30 +27,31 @@
 
 <script>
 import { Sdk } from '@/sdk'
-import { ErrorHandler } from '@/utils/ErrorHandler'
-import config from '@/config'
 import { ApiCallerFactory } from '@/api-caller-factory'
+
+import config from '@/config'
+import { ErrorHandler } from '@/utils/ErrorHandler'
 
 export default {
   props: {
     accountId: {
       type: String,
-      default: ''
+      default: '',
     },
     balanceId: {
       type: String,
-      default: ''
+      default: '',
     },
     isTitled: {
       type: Boolean,
-      default: true
-    }
+      default: true,
+    },
   },
 
   data: _ => ({
     email: '',
     isMasterAccount: false,
-    isLoading: false
+    isLoading: false,
   }),
 
   async created () {
@@ -78,7 +79,7 @@ export default {
           .createCallerInstance()
           .getWithSignature('/identities', {
             filter: { address: accountId },
-            page: { limit: 1 }
+            page: { limit: 1 },
           })
         this.email = ((data || [])[0] || {}).email || this.email
       } catch (error) {
@@ -95,10 +96,7 @@ export default {
       } else {
         return ''
       }
-    }
-  }
+    },
+  },
 }
 </script>
-
-<style scoped>
-</style>