+++ src/components/common/modals/Modal.vue
@@ -1,14 +1,20 @@
 <template>
   <div class="modal" @click="onClick">
-    <div class="modal__content" :style="{ maxWidth: maxWidth, minWidth: minWidth }">
-      <slot/>
+    <div
+      class="modal__content"
+      :style="{ maxWidth: maxWidth, minWidth: minWidth }"
+    >
+      <slot />
     </div>
   </div>
 </template>
 
 <script>
 export default {
-  props: ['maxWidth', 'minWidth'],
+  props: {
+    minWidth: { type: String, default: '' },
+    maxWidth: { type: String, default: '' },
+  },
 
   created () {
     document.addEventListener('keydown', this.onDocumentKeyDown, false)
@@ -44,8 +50,8 @@ export default {
 
     emitCloseRequest () {
       this.$emit('close-request')
-    }
-  }
+    },
+  },
 }
 </script>