+++ src/store/plugins.js
@@ -26,5 +26,7 @@ const localStoragePlugin = store => {
 }
 
 // TODO: setup env
-// export default process.env.NODE_ENV !== 'production' ? [localStoragePlugin] : [localStoragePlugin]
+// export default process.env.NODE_ENV !== 'production'
+//   ? [localStoragePlugin]
+//   : [localStoragePlugin]
 export default [localStoragePlugin]